Change from space to tabs for Caddyfile
Spaces instead of tabs for indentation in Caddyfile will trigger an odd
warning in the log file along with a note on how to correct. Use `caddy
fmt --overwrite` in the same directory as the Caddyfile. Caddy is
available in the Fedora repos so this is feasible even if caddy is not
used on the workstation as a web server per se.