Correct domain syntax with reverse proxy
This instance of forgejo is deployed as a container with a macvlan IP
address along with a caddy based reverse proxy also deployed as a
container on the same macvlan.  The reverse proxy handles https and
connects via http to forgejo. The domain is `forge.lan` and
`git.forge.lan` for ssh access to the git repositories.