Add addition nebula-sync fixes
Nebula-sync can use docker compose secrets as the source for
authentication criteria to pihole instances. There are two secrets
files: 1 for primary, the second for all replicas. The content of the
files are 'url|secret' for primary and 'url|secret,url|secret,...' for
replicas. This is different from pihole which will use docker compose
secrets for the password only.